[Cache] move InMemoryLRUSimpleCache from Adapter namespace to base Cache namespace
